finish restaurant recommendations api calling
